+++ tests/Fixtures/AttributedAggregatesSameName/SnapshotableProfileWithAttribute.php
@@ -5,7 +5,7 @@ namespace Patchlevel\EventSourcingBundle\Tests\Fixtures\AttributedAggregatesSame
 use Patchlevel\EventSourcing\Aggregate\AggregateChanged;
 use Patchlevel\EventSourcing\Aggregate\AggregateRoot;
 use Patchlevel\EventSourcing\Aggregate\SnapshotableAggregateRoot;
-use Patchlevel\EventSourcingBundle\Attributes\Aggregate;
+use Patchlevel\EventSourcingBundle\Attribute\Aggregate;
 
 #[Aggregate(name: 'profileWithAttribute', snapshotStore: 'default')]
 class SnapshotableProfileWithAttribute extends SnapshotableAggregateRoot
