+++ tests/Unit/PatchlevelEventSourcingBundleTest.php
@@ -49,6 +49,7 @@ use Psr\Cache\CacheItemPoolInterface;
 use Psr\SimpleCache\CacheInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\Messenger\MessageBusInterface;
 use InvalidArgumentException;
